refactor: separate `plugin.Store` to its own file
